+++ CHANGELOG.md
@@ -5,5 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New option `--zero-torque` for `bolthumanoid_show_data` to start in
+  zero-torque mode.
 
 Initial version based on `robot_interfaces_solo` v0.9.0-43-g29a0e66